% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiAnalysis-package.R
\docType{package}
\name{packageStarter-package}
\alias{packageStarter}
\alias{packageStarter-package}
\title{packageStarter:}
\description{

}
\author{
\strong{Maintainer}: Etienne Camenen \email{etienne.camenen@gmail.com}

}
\keyword{internal}
