% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageStarter-package.R
\docType{package}
\name{packageStarter-package}
\alias{packageStarter}
\alias{packageStarter-package}
\title{packageStarter: packageStarter}
\description{
packageStarter
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ecamenen/packageStarter}
  \item Report bugs at \url{https://github.com/ecamenen/packageStarter/issues}
}

}
\author{
\strong{Maintainer}: Etienne Camenen \email{etienne.camenen@gmail.com}

}
\keyword{internal}
